fix code duplication, add GenericAvroRecordBinarySerializer and sampler
