+++ src/main/java/net/coru/kloadgen/serializer/GenericAvroRecordBinarySerializer.java
@@ -1,0 +1,43 @@
+/*
+ *  This Source Code Form is subject to the terms of the Mozilla Public
+ *  * License, v. 2.0. If a copy of the MPL was not distributed with this
+ *  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ */
+
+package net.coru.kloadgen.serializer;
+
+import lombok.extern.slf4j.Slf4j;
+import org.apache.avro.generic.GenericRecord;
+import org.apache.avro.io.DatumWriter;
+import org.apache.avro.io.Encoder;
+import org.apache.avro.io.EncoderFactory;
+import org.apache.avro.specific.SpecificDatumWriter;
+import org.apache.kafka.common.header.Headers;
+import org.apache.kafka.common.serialization.Serializer;
+
+import java.io.ByteArrayOutputStream;
+import java.io.IOException;
+
+@Slf4j
+public class GenericAvroRecordBinarySerializer<T extends GenericRecord> implements Serializer<T> {
+
+    @Override
+    public byte[] serialize(String s, T data) {
+        DatumWriter<T> writer = new SpecificDatumWriter<>(data.getSchema());
+        try (final ByteArrayOutputStream baos = new ByteArrayOutputStream()) {
+            Encoder encoder = EncoderFactory.get().binaryEncoder(baos, null);
+            writer.write(data, encoder);
+            encoder.flush();
+            return baos.toByteArray();
+        } catch (IOException e) {
+            log.error("Serialization error for date: {}", data, e);
+            return new byte[]{};
+        }
+    }
+
+    @Override
+    public byte[] serialize(String topic, Headers headers, T data) {
+        return serialize(topic, data);
+    }
+
+}
